restructure directory and add font
