+++ docs/enemy_types.txt
