+++ main.lua
@@ -1,14 +1,19 @@
-local push = require("push")
+local push = require("lib/push")
 local game_width, game_height = 480, 270
 local window_width, window_height = love.window.getDesktopDimensions()
 
+font_mitochondria = love.graphics.newFont('assets/fonts/Mitochondria.ttf', 8)
+font_press_start = love.graphics.newFont('assets/fonts/PressStart2P.ttf', 8)
+love.graphics.setFont(font_press_start)
+
+-- this makes the low res graphics
 love.graphics.setDefaultFilter("nearest", "nearest")
 push:setupScreen(game_width, game_height, window_width, window_height, {fullscreen = true, pixelperfect = true, highdpi = true})
 
 local entities = {}
 local bullets = {}
 
-
+-- this is for triangle man
 function create_triangle_player(x, y)
 	local posX = x or 0
 	local posY = y or 0
@@ -16,12 +21,11 @@ function create_triangle_player(x, y)
 		pos = {x = posX, y = posY},
 		vel = {x = 0, y = 0},
 		max_speed = 100,
-		accel = 750,
-		decel = 2500,
 		facing = 0,
 		attack_speed = 0.1
 	}
 	function player:shoot()
+		-- this is not complete atm but will call attack_speed stat
 		local bullet = {
 			pos = {x = player.pos.x, y = player.pos.y},
 			vel = {x = dx, y = dy},
@@ -36,12 +40,14 @@ function create_triangle_player(x, y)
 		end
 	end
 	function player:update(dt)
+		-- use this in love:update, probably add this to some overarching table like dudes {}
 		player.vel = {x = 0, y = 0}
 		if love.keyboard.isDown('a') then player.vel.x = -1 end
 		if love.keyboard.isDown('d') then player.vel.x = 1 end
 		if love.keyboard.isDown('w') then player.vel.y = -1 end
 		if love.keyboard.isDown('s') then player.vel.y = 1 end
 
+		-- normalize movement here
 		local vel_len = math.sqrt(player.vel.x^2 + player.vel.y^2)
 		print(vel_len)
 		if vel_len > 0 then
@@ -49,16 +55,19 @@ function create_triangle_player(x, y)
 			player.vel.y = player.vel.y / vel_len
 		end
 
+		-- update player position based on velocity table
 		player.pos.x = player.pos.x + player.vel.x * player.max_speed * dt
 		player.pos.y = player.pos.y + player.vel.y * player.max_speed * dt
-
 		player.facing = math.atan2(m_y - player.pos.y, m_x - player.pos.x)
 	end
 	function player:draw()
+
 		local size = 7.5
 		love.graphics.push()
 		love.graphics.translate(player.pos.x, player.pos.y)
 		love.graphics.rotate(player.facing)
+
+		-- i dont know wtf this is but it makes a triangle polygon goated
 		love.graphics.polygon('fill', {
 			size, 0,
 			-size/2, size/2,
@@ -98,12 +107,14 @@ end
 
 
 function love.draw()
+	-- include
 	push:start()
 		player:draw()
+		for i = #bullets, 1, -1 do
+			bullets[i]:draw()
+		end
+		love.graphics.print(player.vel.x, 0, 0)
+		love.graphics.print(player.vel.y, 0, 10)
 	push:finish()
-	for i = #bullets, 1, -1 do
-		bullets[i]:draw()
-	end
-	love.graphics.print(player.vel.x, 0, 0)
-	love.graphics.print(player.vel.y, 0, 20)
+	
 end
